Qtech/search_engine/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.7
django-bootstrap4==2.3.1
django-heroku==0.3.1
factory-boy==3.2.0
Faker==6.6.3
gunicorn==20.1.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
wincertstore==0.2
